+++ drag-drop/src/models/drag-drop.ts
@@ -1,0 +1,11 @@
+// Drag & Drop Interfaces
+export interface Draggable {
+  dragStartHandler(e: DragEvent): void;
+  dragEndHandler(e: DragEvent): void;
+}
+
+export interface DragTarget {
+  dragOverHandler(e: DragEvent): void;
+  dropHandler(e: DragEvent): void;
+  dragLeaveHandler(e: DragEvent): void;
+}
