+++ drag-drop/src/models/project.ts
@@ -1,0 +1,14 @@
+export enum ProjectStatus {
+  Active,
+  Finished,
+}
+
+export class Project {
+  constructor(
+    public id: string,
+    public title: string,
+    public description: string,
+    public people: number,
+    public status: ProjectStatus
+  ) {}
+}
